Metodos XML
son métodos xml de representacion de datos de factura y deben estar en
la clase factura del erp.
